requirements.txt
hikari-lightbulb~=1.2.5